今日收尾

+++ CustomB/Classes/Modules/订单/ViewModel/TLCalculatePriceManager.h
@@ -18,7 +18,7 @@
 //6个元素必须同时配置才能进行计算
 
 @property (nonatomic, assign) float mianLiaoPrice; //单价
-@property (nonatomic, assign) NSInteger mianLiaoCount; //面料数量
+@property (nonatomic, assign) float mianLiaoCount; //面料数量
 @property (nonatomic, assign) float jiaGongPrice; //加工费
 @property (nonatomic, assign) float gongYiPrice; //工艺费
 //
